Analysis, Evidence, Readme, delete some comments

+++ DesignPattern43_Balatro_Pretest/src/Modifiers/IModifier.h
@@ -3,7 +3,7 @@
 class IModifier {
 public:
   // Fungsi ini wajib dimiliki oleh semua wujud nyata kartu modifier nanti
-  virtual int ApplyModifier(int currentScore) = 0;
+  virtual long long ApplyModifier(long long currentScore) = 0;
 
   // Ini adalah pelengkap wajib di C++ agar memori aman saat menghapus objek
   virtual ~IModifier() {}
