+++ DesignPattern43_Balatro_Pretest/src/Modifiers/TripleScoreModifier.h
@@ -3,5 +3,5 @@
 
 class TripleScoreModifier : public IModifier {
 public:
-  int ApplyModifier(int currentScore) override;
+  long long ApplyModifier(long long currentScore) override;
 };
